mutt_compose_editor safe .swp file storage
mutt puts its message body file in /tmp, e.g.
/tmp/mutt-brachium-1000-6813-470748004361105733 .

Normally, if the computer reboots, this file and the vim swap file will
both be lost in distros that clear /tmp on boot.  This is because the
default setting is directory=.,~/tmp,/var/tmp,/tmp (which is a list of
candidate directories in which to write the swap file, with "."
representing the location of the file being edited rather than vim's
current directory).

This fix to mutt_compose_editor sets directory=~/tmp which means the
swap file will still be around after a reboot.